requirements.txt
langchain
openai
faiss-cpu
streamlit
streamlit-chat
google-api-python-client
google-auth-httplib2
google-auth-oauthlib